rename xpkg.Resolver to xpkg.Marshaler
