+++ internal/xpkg/dep/marshaler/xpkg/testdata/config_package.yaml
