+++ internal/xpkg/dep/marshaler/xpkg/testdata/crossplane.yaml
