+++ internal/xpkg/dep/marshaler/xpkg/testdata/provider_meta.yaml
