+++ internal/xpkg/dep/marshaler/xpkg/testdata/provider_package.yaml
