+++ internal/xpkg/dep/marshaler/xpkg/testdata/providerconfigs.helm.crossplane.io.yaml
